Setup initial 16x16 grid using javascript function
I've been following along a YouTube tutorial ("Code with Rob") to help
me complete the etch-a-sketch project setup. So far we have created the
index.html outline for the initial board and button areas, and
also began using the css grid method to style the board space.
Lastly, we've written a short javascript function to populate the
etch-a-sketch board area with 16x16 grid, which can be variably
expanded later on the user's input for desired grid size.