Add color sketch functionality and ability to resize sketch board
Modified javascript code to allow for various colors to be added
during mouseover events across the etch-a-sketch board. Also added
the ability for user to resize the number of board rows/columns to
provide higher/lower levels of granularity, limited from 2-100 rows
or columns. Lastly, added ability to switch between different pen
colors, including a rainbow option, and white color as an "eraser".